notify: [windows] consistent semantics
The application callback will receive the ID of the parent notification
as the first string argument.

All the extra data, including paylods and whatever action was invoked will
appear in the userData map.

Values are (hex) encoded before submitting the notification so that we
can avoid any complex characater escaping logic.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>